# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Multi-Cloud Terrarium (mc-terrarium) aims to provide an environment to enrich multi-cloud infrastructure.",
        "title": "Multi-Cloud Terrarium REST API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/cloud-barista/mc-terrarium/issues/new"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "v0.1.2"
    },
    "host": "localhost:8055",
    "basePath": "/terrarium",
    "paths": {
...
        "/tr/{trId}/testbed/actions/apply": {
            "post": {
                "description": "Apply the testbed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Testbed] OpenTofu Actions (for fine-grained control)"
                ],
                "summary": "Apply the testbed",
                "parameters": [
                    {
                        "type": "string",
                        "default": "testbed01",
                        "description": "Terrarium ID",
                        "name": "trId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Custom request ID",
                        "name": "x-request-id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Credential holder (profile) name",
                        "name": "X-Credential-Holder",
                        "in": "header"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/tr/{trId}/testbed/actions/destroy": {
            "delete": {
                "description": "Destroy the testbed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Testbed] OpenTofu Actions (for fine-grained control)"
                ],
                "summary": "Destroy the testbed",
                "parameters": [
                    {
                        "type": "string",
                        "default": "testbed01",
                        "description": "Terrarium ID",
                        "name": "trId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Custom request ID",
                        "name": "x-request-id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Credential holder (profile) name",
                        "name": "X-Credential-Holder",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/tr/{trId}/testbed/actions/emptyout": {
            "delete": {
                "description": "EmptyOut the testbed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Testbed] OpenTofu Actions (for fine-grained control)"
                ],
                "summary": "EmptyOut the testbed",
                "parameters": [
                    {
                        "type": "string",
                        "default": "testbed01",
                        "description": "Terrarium ID",
                        "name": "trId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Custom request ID",
                        "name": "x-request-id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Credential holder (profile) name",
                        "name": "X-Credential-Holder",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/tr/{trId}/testbed/actions/init": {
            "post": {
                "description": "Init testbed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Testbed] OpenTofu Actions (for fine-grained control)"
                ],
                "summary": "Init testbed",
                "parameters": [
                    {
                        "type": "string",
                        "default": "testbed01",
                        "description": "Terrarium ID",
                        "name": "trId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Parameters requied to create a testbed",
                        "name": "ReqBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateTestbedRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Custom request ID",
                        "name": "x-request-id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Credential holder (profile) name",
                        "name": "X-Credential-Holder",
                        "in": "header"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/tr/{trId}/testbed/actions/output": {
            "get": {
                "description": "Output the testbed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Testbed] OpenTofu Actions (for fine-grained control)"
                ],
                "summary": "Output the testbed",
                "parameters": [
                    {
                        "type": "string",
                        "default": "testbed01",
                        "description": "Terrarium ID",
                        "name": "trId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "refined",
                            "raw"
                        ],
                        "type": "string",
                        "default": "refined",
                        "description": "Resource info by detail (refined, raw)",
                        "name": "detail",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": true,
                        "description": "Refresh the state before getting the info",
                        "name": "refresh",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Custom request ID",
                        "name": "x-request-id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Credential holder (profile) name",
                        "name": "X-Credential-Holder",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CreateTestbedRequest": {
            "type": "object",
            "properties": {
                "testbed_config": {
                    "$ref": "#/definitions/model.TestbedConfigDetail"
                }
            }
        },
...
        "model.Response": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string",
                    "example": "Any details"
                },
                "list": {
                    "type": "array",
                    "items": {}
                },
                "message": {
                    "type": "string",
                    "example": "Any message"
                },
                "object": {
                    "type": "object",
                    "additionalProperties": true
                },
                "status": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "model.TestbedConfigDetail": {
            "type": "object",
            "properties": {
                "desired_providers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "aws",
                        "azure",
                        "gcp",
                        "alibaba",
                        "tencent",
                        "ibm",
                        "dcs"
                    ]
                },
                "terrarium_id": {
                    "type": "string",
                    "example": ""
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}